More fixes for PHPTAL 1.0 on PHP5. Apparently it's a bug in PHPTAl 0.7 that substitutions with ${name} are not HTML-escaped... Replaced various inline bits with tal:content or tal:replace using the 'structure' keyword to do this in a compatible way, but some bits may be remaining.
Also changed some uses to tal:attributes, and took the htmlspecialchars()
off of the settings in SkinPHPTal to let them get escaped normally.
The main page logo link is currently bitten by this as i18n:attributes
and tal:attributes can't be used together on 0.7.0; this needs tweaking.

PHPTAL 1.0 currently doesn't escape the tal:attributes bits, but this seems
to be a simple bug and should be fixed soon, I hope.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@5557 dd0e9695-b195-4be7-bd10-2dea1a65a6b6