Committing one of my working copies on a branch for reference.
Adapts the $wgNamespaces array from wikidata branch to an iteratable
object that looks and feels like an array, but returns a usable object
if you specifically ask for a bogus namespace. This will allow the
system to gracefully recover from bad namespace values recorded in
the database. Requires PHP 5.

This branch has cropped out most of the database backing for the
namespaces, to concentrate on the API work.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/branches/namespace-testing/phase3@13751 dd0e9695-b195-4be7-bd10-2dea1a65a6b6