New special page Special:Deletedcontribs for browsing deleted edits made by a user.
NOTENOTENOTE: This is a one-night hack; feel free to tweak it mercilessly.  It has
a lot of common code with Special:Contributions that should be factored out.  There
is no link to this new page anywhere in the UI yet; I'd like people to test and
review this a bit before officially announcing it.  The page is currently only
available to users with deletedhistory privileges; a possible additional feature
might be to allow all users to list (but not view) their _own_ deleted edits.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@17968 dd0e9695-b195-4be7-bd10-2dea1a65a6b6