* (bug 1438) Fix for diff table layout on very wide lines for Gecko and   Opera-based browsers (incomplete, does not fix KHTML or MSIE)
This adds a <div> inside the <td> for diff content cells, and declares
the style for these <div>s to have overflow: auto.

In Gecko (tested Firefox 2.0) and Opera (tested 9.10 and over) this does
two things:
* The table layout treats the cells as the requested width instead of bloating
  out to the widest line of content, so the table stays visible on screen
* The individual cells that are too long get horizontal scroll bars

Unfortunately this doesn't have the hoped-for effects in other tested browsers:
* MSIE 7
* Safari 2
* Konqueror 3.5
* iCab 3

But neither has it any ill effects, so... it's a start.
There's probably some other way to force the layout algorithm to behave
that I haven't quite stumbled on yet... Might have better luck with the
fixed table layout option, though that seems less friendly to the little
+ and - columns.

The C++ diff plugins will have to be updated to support this scheme, but
no harm will be done if they're not (just they won't do anything new).


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@22192 dd0e9695-b195-4be7-bd10-2dea1a65a6b6