Tweaks to the AJAX license preview: * Rename 'licence' to 'license' for consistency with the rest of the code. * The preview code was assuming simple templates with no parameters and no subst:s, which failed on most of the selections used now on Commons. Now doing a full pre-save transform and parse on {{$license}}, just as will be done on the final save. * When selecting 'None' again after another option, the preview pane is now cleared.
A fun 'todo' might be to also pass the currently selected filename, if any, to the license preview. Some of the templates in use attempt to use the current-page-name variables to include a self-link. At the moment using a hardcoded 'Image:Sample.jpg' for the virtual title on the preview rendering.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@24247 dd0e9695-b195-4be7-bd10-2dea1a65a6b6