Revert r30162 for now.
A problem with this is that it *un*watches a previously watched image under the following
circumstances:

* 'watch pages I edit' is not enabled (eg, default state)
* go to Special:Upload and select the file
* hit 'upload'

The initial check state is unchecked (since there was no initial destination name set), and this doesn't get updated to reflect the existing watch state of the previous image.


There are a couple possible ways around this. One is to compare the form's actual initial check state with the submitted check state and only apply an unwatch if there was a difference.

Another might be to do a watch state update via AJAX when a new destination filename is set in the form. This would allow the checkmark's default state to be set 'properly' for those with JS enabled in modern browsers.

Perhaps a combination should be used.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@30163 dd0e9695-b195-4be7-bd10-2dea1a65a6b6