Revert r37078, 37132 "New hook ParserBeforeTranscludeTemplate: This hook allows for modification of the title and text of a template which is being transcluded.
Use of this hook will allow extensions to create features such as TransWiki for an alternative to ScaryTransclusions."

This hook seems a bit oddly placed to me; the template gets fetched locally, and *then* we give the opportunity to fetch it remotely instead? Just seems to be in the wrong order, and pretty unclear.




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@37148 dd0e9695-b195-4be7-bd10-2dea1a65a6b6