(bug 15928) Make special pages dropdown inline
There was a weird extra "|" at the end of lines in Nostalgia (probably other non-MonoBook skins too).  Upon investigation, I found that the special pages dropdown (which immediately followed that separator but on the next line) had a class="inline", which actually did nothing but whose intent was fairly obvious.  I removed the presentational class and just styled it using the id, so now the separator actually . . . separates.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@41940 dd0e9695-b195-4be7-bd10-2dea1a65a6b6