* (bug 11533) Fixed insane slowdown when in read-only mode for long periods   of time with CACHE_NONE (default objectcache table configuration).
MediaWikiBagOStuff was refusing to perform any cache updates when in read-only
mode. This made sense originally when read-only mode was an extreme hardcore
"nothing can write to the DB EVAR because we're doing weird things to it",
but today we do allow backend writes of various sorts, and it's mainly meant
to stop user-level activity.

The refusal to perform updates caused the MessageCache to spend a long period
of time retrying the storage of the cache if it had expired out of the table previously... which is kind of bad as every page load would be insanely slow. :D


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@42796 dd0e9695-b195-4be7-bd10-2dea1a65a6b6