Abstract more methods in DatabaseBase
Notably, this will switch conditional() in MySQL from using IF() to
using CASE, like all other DBMSes.  Documentation suggests this works
back to 4.0.  If it's a problem, it's a matter of a few lines to
override it in DatabaseMysql.php.

Also, some extra explanatory comments have been added to a number of
methods in DatabaseBase.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@52002 dd0e9695-b195-4be7-bd10-2dea1a65a6b6