Backport r49655, fatal errors for some users
I have not tested that this is safe, beyond php -l.  However, ZeiP on
IRC reported that it fixed his fatals errors and seemed to work.

Original summary:

Move checkRedirect() from FileRepo to LocalRepo, leaving only a stub
behind. It's inappropriate for FileRepo to be accessing the database or
memcached, this is exclusively a job for a subclass. Avoids error "Call
to undefined method <class>::getArticleID()" whenever anyone uses a repo
which is not derived from LocalRepo, such as FSRepo or ForeignAPIRepo.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/branches/REL1_15/phase3@52009 dd0e9695-b195-4be7-bd10-2dea1a65a6b6