* (bug 20364) Fixed regression in GIF metadata loading
There was a missing parameter in GIFMetadataExtractor's skipBlock() call for 'netscape 2.0' data blocks, which threw a monkey in the works.
Now also checking for exceptions thrown by the metadata load and stubbing out null metadata for files which can't be read, rather than letting the exception bubble up and kill MediaWiki. :)


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@55541 dd0e9695-b195-4be7-bd10-2dea1a65a6b6