maxlength=200 for page move summary in HTML5
Bug 16921.  maxlength is not allowed on textareas in HTML4, so this only
works in HTML5.  Note that Firefox 3.5 and Opera 9.22 ignore the
attribute (didn't test IE), so this isn't a complete fix.  Recent WebKit
does respect the attribute (tested in Chrome 4).

Of course, the length limit of 200 is a hack, just like for edit
summaries, and we really need to move to a non-varchar(255) backend for
all these fields.

Relevant to r45517, r45571.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@60054 dd0e9695-b195-4be7-bd10-2dea1a65a6b6