Fix IEFixes.js loading again
Bug 25167.  The fix in r65774 was wrong -- apparently I didn't test well
enough.  I've tested the new fix by loading IE6 and requesting a page
from my wiki without it, verifying with Wireshark that no HTTP request
was sent for IEFixes.js, and then repeating with the fix applied to
verify it works.

Will backport to 1.16, like r65774, since this is still a regression
from 1.15.  Note that only IE6 and less were ever affected, and things
seem to be mostly usable without the fixes file, so I don't think this
is of enormous importance.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@73978 dd0e9695-b195-4be7-bd10-2dea1a65a6b6