* (bug 27333) Fix repetitive last-seen time queries on page history
Title::getNotificationTimestamp() is called for every line in HistoryPager, to compare against the revision's timestamp to determine if it's a new edit since the viewer's last visit to the page.
The function kept an internal cache within the Title object to avoid having to make the query over and over in this situation, but in the case that it returns null, the cached entry would be accidentally ignored on the next round due to use of isset() to check for the array entry. (Most of the time it's great, but if the value of the key actually *is* null, isset() returns false.) Switched to array_key_exists(), which doesn't have this false negative. Now only one query gets issued for this on a history page, even if the answer is null.




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@82004 dd0e9695-b195-4be7-bd10-2dea1a65a6b6