(bug 27854) Http::isValidURI is way to lax. This is a much simplified regex that accepts a subset of the previous regex, but also accepts ftps because both cURL and php support it. It no longer accepts thing like 'foo http://bar bax' which was my main concern
Note the previous regex kind of looks more restrictive, but is not since saying "anything not containing a space
optionally followed by anything not containing a bunch of characters including a space" is the same as saying anything
with no spaces. See also r83296. This obviously doesn't catch all cases, but I personally think its sufficient.
At the very least it is a very significant improvement over the previous version that caught almost nothing.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@83360 dd0e9695-b195-4be7-bd10-2dea1a65a6b6