Backport r83716 "Normalize named entities to numeric"
Didn't bother merging the parser tests, so those are only in trunk.
Original commit message:

We should never be outputting named entities other than the ones in XML,
&lt; &gt; &amp; &quot;, because that will break well-formedness unless
we have a DTD in the doctype, which we don't in HTML5 mode.

I stuck with outputting numeric entities here instead of UTF-8 because
some characters are hard to read in UTF-8 (e.g., &nbsp;).  Maybe it
would be nicer if we decoded to UTF-8 except for whitespace and control
characters, or something like that, but it's a detail.

I'll backport to 1.17 and add RELEASE-NOTES there, which is why I added
the line to HISTORY instead of RELEASE-NOTES.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/branches/REL1_17/phase3@83717 dd0e9695-b195-4be7-bd10-2dea1a65a6b6