Backport r83061 "Disable $wgExperimentalHtmlIds again by default"
No RELEASE-NOTES because it's only a regression fix, it was disabled by
default in 1.16.  Removed RELEASE-NOTES for r70526.  Original commit
message:

See bug 27733, bug 27694, bug 27474.  Reverts r70526.  Should be
backported to 1.17 (along with removing r70526's RELEASE-NOTES).  Can be
ported to WMF as well, but it's kind of trivial to replicate in
LocalSettings.php, so doesn't really have to be.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/branches/REL1_17/phase3@83720 dd0e9695-b195-4be7-bd10-2dea1a65a6b6