* (bug 26937) [Installer] Fix for Javascript-opened sections being incorrectly open or closed when returning to page
The sections whose visibility was being controlled by the checkbox had an initial visibility hardcoded in the page output, which had been manually set to match the expected initial default state. I've switched these to instead check what the variable is set to and base the initial style visibility on that.
Could be better to encapsulate these settings together in some way, but this'll do for now.

Note that the way these are all done now means that the sections can never be used when JS is off; if a fully-functional installer is desired in a JS-free environment, it may be necessary to change these further to default to shown, and do the initial hiding in JavaScript (could have 'flashing' issues though: the bits showing up, then quickly hiding. laaame but oh well)




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@84875 dd0e9695-b195-4be7-bd10-2dea1a65a6b6