Revert r77555 and followups r77563, r77572, r78116 (merge DisableAccount to core).
Per CR at the time: this creates a nearly irreversable action that is not nearly well documented enough (even if disabled by default). 

We already have the $wgBlockDisablesLogin kludge in place for this. If we're going to do more work on this idea, it should be well thought out, not another hack.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@86146 dd0e9695-b195-4be7-bd10-2dea1a65a6b6