Remove the JPEG/TIFF specific metadata code from BitmapHandler and put it in JpegOrTiff handler to stop mostly irrelevent classes from getting it.
Also remove a method that is an exact duplicate of a base class (not sure whats with that).

This also coincidently fixes the issue with when a foreign file repo uses PagedTiffHandler
and the local one does not, and the builtin Tiff handler tries to treat the metadata as if
it was its own form.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@86567 dd0e9695-b195-4be7-bd10-2dea1a65a6b6