Makes sure wgAdditionalMailParams is null in safe mode
The global is used for a call to PHP mail() function as a way to add
additional parameters.  This will cause mail() to send E_NOTICE when
using safe mode.

Since we could use the global at different places, it makes sens to
ensure it has a sane value through Setup.php

Follow up r75557



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@88391 dd0e9695-b195-4be7-bd10-2dea1a65a6b6