(follow-up to some RequestContext revision) When Special:userlogin creates a new user, not only replace $wgUser, but also do RequestContext::getMain()->setUser( $new_user );
Otherwise this causes a bug in the page immediatly after account creation where the
personal url's go to the wrong place, and the js wgUser variables are wrong.

I didn't add any release notes, since I'm tagging this for 1.18, and it was broken in 1.18,
so its not fixing anything between versions. (hopefully that's correct thing to do).


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@90934 dd0e9695-b195-4be7-bd10-2dea1a65a6b6