(bug 28762) Resizing to specified height broken for very thin images
Restructure and comment of ImageHandler::normaliseParams(). ImageHandler::normaliseParams() will now output both width/height as physicalWidth/Height, where width/height are the dimensions used by the browser, and physicalWidth/Height are the dimensions of the thumbnail that will be created. This allows keeping the use of unique dimensions for a specified thumbnail width, by forcing client side resizing of 1px-wide images with non corresponding height.
Also fixed a bug where the check for mustRender() was in an if statement where it should not have been.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@92279 dd0e9695-b195-4be7-bd10-2dea1a65a6b6