* (bug 31282) Fix use of ForeignAPIRepo/InstantCommons TIFF images when $wgTiffThumbnailType is left unset.
When pulling files over ForeignAPIRepo, the source site is what creates the thumbnails; we don't care whether we can locally produce thumbs of TIFFs.
This feels a bit hacky, and may indicate we need to better refactor media handlers & file repositories. As long as the source site can produce a thumbnail image URL or a chunk of HTML to export, the local site shouldn't have to touch any of that stuff and should just be passing data through.




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@98812 dd0e9695-b195-4be7-bd10-2dea1a65a6b6