Give a clear error message instead of un-intelligible UNIQ.*QINU markers.
Not sure the preg_match() is actually needed.  Or it may be
appropriate to use MARKER_SUFFIX for the match.

The error message may also need to be rewritten to be more
user-friendly, but I'm pretty sure *an* error message is friendlier
than UNIQ garbage.  And making them visible error messages makes them
easier to be found.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@100262 dd0e9695-b195-4be7-bd10-2dea1a65a6b6