QUnit test cases for bug 31847: will trigger a fail on IE 6/7/8
jquery.textSelection's getCaretPosition() gives bogus results on IE 6/7/8 when there's not a selection set in the textarea; it seems to give us some value somewhere in the middle of the textarea, rather than what we expected.
If we actually set an empty selection, then it gets it back just fine!
But if the cursor is in initial state, or has been moved by the user, we get back garbage. This is causing WikiEditor to sometimes insert things in the wrong place, as it tries to pull the caret position and manipulate the selection to do insertions; we then end up using the bogus selection when it gets saved in via selection restoration.





git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@100391 dd0e9695-b195-4be7-bd10-2dea1a65a6b6