Followup r100391: fix caret position test case on Opera, Firefox 3.5/3.6/4.0/5.0
Opera (through 11.52 at least) & Firefox prior to 6.0 set the caret initially to the end of the textarea, which gave a mismatch. All we really need is to know we got a sane value, so checking for both (start == 0; end == text.length) gets the job done.
Under ideal circumstances, we'd check after a user clicks in the textarea on the first line, but no good way to trigger that. :)




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@100656 dd0e9695-b195-4be7-bd10-2dea1a65a6b6