* (bug 31878, bug 31542) Fix XML namespace output in API; removed now-unneeded includexmlnamespace parameter.
Any existing uses of includexmlnamespace should be ignored and harmless.

If output data already includes a specific default namespace as an xmlns attribute, it is now retained -- so for example the RSD API discovery module outputs the appropriate namespace instead of the generic API one.

Partial revert of r99135 which added the includexmlnamespace parameter as a temporary requirement to get the xmlns attribute included.

Followup to r88007 which added the xmlns originally, but in the wrong order so it overrode existing output data.




git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@101912 dd0e9695-b195-4be7-bd10-2dea1a65a6b6