* (bug 32548) fix minification bug when numeric literal with exponent was split over lines
This broke the OpenLayers support in the Maps extension, as used for example on TranslateWiki.net.
The original JavaScriptMinifier's tokenizer (r83885) explicitly didn't bother looking for the exponent part because it "didn't matter" to its internal state machine; however since r83891 added a max line length that definitely is not true.

I've split out handling of hex and decimal numerals, and let the decimal numeral handling check for exponents.

PHPUnit test cases were added in r103846.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk/phase3@103865 dd0e9695-b195-4be7-bd10-2dea1a65a6b6